Added Employment model which is mapping between Profile and Company and the compiled files of the models
