+++ authentication_app/models/__init__.py
@@ -1,1 +1,2 @@
 from authentication_app.models.custom_user import CustomUser
+from authentication_app.models.employment import Employment
